Revert "offline indexedDB"
