+++ app/components/ExpensesForm.js
@@ -1,4 +1,3 @@
-// bm-day-book/app/components/ExpensesForm.js
 "use client";
 
 import { useState } from "react";
