--- app/hooks/useServiceWorker.js
@@ -1,21 +1,0 @@
-"use client";
-// hooks/useServiceWorker.js
-import { useEffect } from "react";
-
-export default function useServiceWorker() {
-  useEffect(() => {
-    if ("serviceWorker" in navigator) {
-      navigator.serviceWorker
-        .register("/sw.js")
-        .then((registration) => {
-          console.log(
-            "Service Worker registered with scope:",
-            registration.scope
-          );
-        })
-        .catch((error) => {
-          console.error("Service Worker registration failed:", error);
-        });
-    }
-  }, []);
-}
