+++ package.json
@@ -12,7 +12,6 @@
     "@tanstack/react-query": "^5.59.16",
     "axios": "^1.7.7",
     "chart.js": "^4.4.6",
-    "idb": "^8.0.0",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.454.0",
     "mongoose": "^8.8.0",
