--- public/service-worker.js
@@ -1,20 +1,0 @@
-// public/service-worker.js
-self.addEventListener("install", (event) => {
-  event.waitUntil(
-    caches.open("static-v1").then((cache) => {
-      return cache.addAll([
-        "/",
-        "/sale-entries",
-        "/path/to/other/static/files",
-      ]);
-    })
-  );
-});
-
-self.addEventListener("fetch", (event) => {
-  event.respondWith(
-    caches.match(event.request).then((response) => {
-      return response || fetch(event.request);
-    })
-  );
-});
