--- public/sw.js
@@ -1,8 +1,0 @@
-// public/sw.js
-self.addEventListener("install", (event) => {
-  console.log("Service Worker installing.");
-});
-
-self.addEventListener("activate", (event) => {
-  console.log("Service Worker activating.");
-});
